# QUICKSTEP-57 Support for performing partitioned aggregation.