… storage/CMakeLists.txt
… storage/StorageBlock.cpp
… storage/StorageBlock.hpp
… utility/CMakeLists.txt
… utility/CompositeHash.hpp