@hakanmemisoglu 2016-10-23
@hbdeshmukh, It seems OK to me. 
I guess Cmake just complains about a circular dependency.
You can remove `quickstep_storage_PartitionedHashTable` library from the CMakeList definition to break circular dependency.