@hbdeshmukh 2016-10-23
Thanks @hakanmemisoglu 

I think the circular dependency is because of the inclusion of PartitionedHashTablePool in Storage Manager. If I cut that link, it should be fine. I may have to rewrite a function in StorageManager, which shouldn't be difficult.