@hbdeshmukh 2016-10-26
Addressed the review comments and re-ran the queries. Here are the results: 

<table class="table table-bordered table-hover table-condensed">
<thead><tr><th title="Field #1">Queries</th>
<th title="Field #2">master oct 26</th>
<th title="Field #3">part-agg oct 26</th>
<th title="Field #4">Speedup over master</th>
</tr></thead>
<tbody><tr><td>1</td>
<td>16,904</td>
<td>16,295</td>
<td align="right">1.04</td>
</tr>
<tr><td>2</td>
<td>766</td>
<td>612</td>
<td align="right">1.25</td>
</tr>
<tr><td>3</td>
<td>5,060</td>
<td>5,105</td>
<td align="right">0.99</td>
</tr>
<tr><td>4</td>
<td>2,332</td>
<td>2,355</td>
<td align="right">0.99</td>
</tr>
<tr><td>5</td>
<td>4,988</td>
<td>5,025</td>
<td align="right">0.99</td>
</tr>
<tr><td>6</td>
<td>402</td>
<td>401</td>
<td align="right">1.00</td>
</tr>
<tr><td>7</td>
<td>28,398</td>
<td>24,098</td>
<td align="right">1.18</td>
</tr>
<tr><td>8</td>
<td>1,818</td>
<td>1,835</td>
<td align="right">0.99</td>
</tr>
<tr><td>9</td>
<td>8,858</td>
<td>12,839</td>
<td align="right">0.69</td>
</tr>
<tr><td>10</td>
<td>13,388</td>
<td>8,126</td>
<td align="right">1.65</td>
</tr>
<tr><td>11</td>
<td>2,321</td>
<td>998</td>
<td align="right">2.32</td>
</tr>
<tr><td>12</td>
<td>2,606</td>
<td>2,531</td>
<td align="right">1.03</td>
</tr>
<tr><td>13</td>
<td>34,860</td>
<td>17,832</td>
<td align="right">1.95</td>
</tr>
<tr><td>14</td>
<td>818</td>
<td>813</td>
<td align="right">1.01</td>
</tr>
<tr><td>15</td>
<td>4,656</td>
<td>900</td>
<td align="right">5.17</td>
</tr>
<tr><td>16</td>
<td>9,411</td>
<td>8,574</td>
<td align="right">1.10</td>
</tr>
<tr><td>17</td>
<td>1,693</td>
<td>1,732</td>
<td align="right">0.98</td>
</tr>
<tr><td>18</td>
<td>92,072</td>
<td>27,665</td>
<td align="right">3.33</td>
</tr>
<tr><td>19</td>
<td>1,430</td>
<td>1,473</td>
<td align="right">0.97</td>
</tr>
<tr><td>20</td>
<td>1,594</td>
<td>1,947</td>
<td align="right">0.82</td>
</tr>
<tr><td>21</td>
<td>139,356</td>
<td>178,787</td>
<td align="right">0.78</td>
</tr>
<tr><td>22</td>
<td>6,895</td>
<td>6,905</td>
<td align="right">1.00</td>
</tr>
<tr><td>Total</td>
<td>380,625</td>
<td>326,850</td>
<td align="right">1.16</td>
</tr>
</tbody></table>