@jianqiao 2016-10-26
LGTM! Merging.